+++ src/components/WindowManager.tsx
@@ -1,6 +1,6 @@
 import { ThemeProvider } from "emotion-theming";
 import { defaultsDeep } from "lodash";
-import React, { PropsWithChildren } from "react";
+import React, { DetailedHTMLProps, HTMLAttributes } from "react";
 import { AppTheme } from "../AppTheme";
 import { WindowManagerContextProvider } from "../context";
 import { ContentGetter } from "./window/WindowContent";
@@ -8,17 +8,15 @@ import { WindowsContainer } from "./WindowsContainer";
 
 const defaultTheme = { backgroundOpacity: 0.9, backdropFilter: "blur(16px)" };
 
-export function WindowManager({
-  children,
-  theme,
-  contentGetter,
-}: PropsWithChildren<{
+interface WindowManagerProps extends DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement> {
   theme: AppTheme;
   contentGetter: ContentGetter;
-}>): JSX.Element {
+}
+
+export function WindowManager({ theme, contentGetter, children, ...props }: WindowManagerProps): JSX.Element {
   return (
     <WindowManagerContextProvider>
-      <div>
+      <div {...props}>
         {children}
         <ThemeProvider<AppTheme> theme={(outerTheme = {}) => defaultsDeep(theme, outerTheme, defaultTheme)}>
           <WindowsContainer contentGetter={contentGetter} />